[Data] modify uint8 to parse sql 'INT(8) UNSIGNED' format to 'INT8 UNSIGNED'

+++ test-suite/src/data_type/uint8.rs
@@ -6,8 +6,8 @@ use {
 test_case!(uint8, async move {
     run!(
         "CREATE TABLE Item (
-            field_one INT(8) UNSIGNED,
-            field_two INT(8) UNSIGNED,
+            field_one INT8 UNSIGNED,
+            field_two INT8 UNSIGNED,
         );"
     );
     run!(r#"INSERT INTO Item VALUES (1, 2), (1, 3), (2, 4), (2, 5);"#);
